法线积分 以及球的相交计算

+++ src/Camera/Image.hpp
@@ -9,7 +9,7 @@ class Image{
 
     vec3 getPixel(int x,int y) const ;
 
-    size_t getIndex(int x,int y) const ;
+    size_t getIndex(size_t x,size_t y) const ;
 
     std::string  outputFileName;
     std::vector<Float> pixels;
